Big refactor around to reduce complexity and length.
Refactored SyntaxTrieFactory to clean the code and use pure DataObjects.

Refactored SyntaxTrieFactory tests in two files and removed some lower value tests to pass Rubocop's length/complexity
